# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiMessageResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PaginationMeta": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ProductResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image_url": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"price",
					"image_url",
					"status",
					"category",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductListResponse": {
				"properties": {
					"products": {
						"items": {
							"$ref": "#/components/schemas/ProductResponse"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"products",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ProductListResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ProductListResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ProductResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ProductResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ProductResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ProductResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatedProductResponse": {
				"properties": {
					"productId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"productId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CreatedProductResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CreatedProductResponse"
					}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProductBody": {
				"properties": {
					"category_id": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"name": {
						"type": "string",
						"example": "iPhone 15"
					},
					"description": {
						"type": "string",
						"example": "Latest Apple smartphone"
					},
					"price": {
						"type": "number",
						"format": "double",
						"example": 999.99
					},
					"image_url": {
						"type": "string",
						"example": "https://example.com/iphone.png"
					}
				},
				"required": [
					"name",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductBody": {
				"properties": {
					"category_id": {
						"type": "number",
						"format": "double",
						"example": 2
					},
					"name": {
						"type": "string",
						"example": "iPhone 15 Pro"
					},
					"description": {
						"type": "string",
						"example": "Updated description"
					},
					"price": {
						"type": "number",
						"format": "double",
						"example": 1199.99
					},
					"image_url": {
						"type": "string",
						"example": "https://example.com/new.png"
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						],
						"example": "active"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "BMD Training API",
		"version": "1.0.0",
		"description": "API documentation for BMD Training backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/admin/products": {
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_ProductListResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get all products",
				"tags": [
					"Admin - Products"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_CreatedProductResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Create a new product",
				"tags": [
					"Admin - Products"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductBody"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Update a product",
				"tags": [
					"Admin - Products"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Delete a product",
				"tags": [
					"Admin - Products"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/products/category": {
			"get": {
				"operationId": "GetProductsByCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_ProductResponse-Array_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get products by category",
				"tags": [
					"Admin - Products"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/products/{productId}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiResponse_ProductResponse_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Get a product by ID",
				"tags": [
					"Admin - Products"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/admin/products/restore": {
			"put": {
				"operationId": "RestoreProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiMessageResponse"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"summary": "Restore a deleted product",
				"tags": [
					"Admin - Products"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}